moved reused styled componenets to own module
