+++ src/renderer/components/ClusterPage/clusterPageComponents.js
@@ -1,0 +1,25 @@
+import styled from '@emotion/styled';
+import { layout } from '../styles';
+
+export const DrawerTitleWrapper = styled.div(() => ({
+  paddingLeft: layout.pad * 3,
+  paddingRight: layout.pad * 3,
+  marginTop: -layout.pad * 3,
+  marginBottom: -layout.pad * 3,
+}));
+
+export const DrawerItemsWrapper = styled.div(() => ({
+  paddingLeft: layout.pad * 3,
+  paddingRight: layout.pad * 3,
+  paddingBottom: layout.pad * 2.25,
+  backgroundColor: 'var(--contentColor)',
+
+  '& > div': {
+    paddingTop: layout.pad * 1.5,
+    paddingBottom: layout.pad * 1.5,
+  },
+}));
+
+export const Link = styled.a(() => ({
+  color: 'var(--primary)',
+}));
